Separate complete safe-value policy from error presentation
